add clear route and change ts to u64

+++ src/api/requests.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 pub struct UpdatePayload {
-    pub ts: usize,
+    pub ts: u64,
     pub arm_id: usize,
     pub reward: f64,
 }
